requirement.txt
newsapi-python==0.2.5
tinydb==3.13.0
feedparser==5.2.1
twitter-twitter_scraper==0.3.0
ciso8601==2.1.1
html-purifier==0.1.9
